refactor(overview): refactor feed filters to sort options
