--- lib/views/filter_widgets/timeline_filter/timeline_filters.dart
@@ -1,13 +1,0 @@
-/// This enum is used to create the timeline filters.
-/// It contains the name of the filters.
-enum TimeLineFilters {
-  nearest("Nearest"),
-  best("Best"),
-  newFilter("New"),
-  top("Top"),
-  hot("Hot");
-
-  final String name;
-
-  const TimeLineFilters(this.name);
-}
